v0.0.6
+ middleware support in bot server & refactoring
+ custom logger (todo: refactor dependency logic)

+++ backend/internal/service/v1/user.go
@@ -15,7 +15,7 @@ func NewUserService(repo repository.UserRepository) *UserService {
 	return &UserService{repo: repo}
 }
 
-func (s *UserService) GetByTgID(ctx context.Context, tgID uint) (*models.User, error) {
+func (s *UserService) GetByTgID(ctx context.Context, tgID int64) (*models.User, error) {
 	if tgID == 0 {
 		return nil, TgIdRequired
 	}
